requirements.dev.txt
flake8==6.1.0
django-debug-toolbar==4.2.0